requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
google-genai>=1.24.0
aiomysql>=0.2.0
psutil
trafilatura